Replication and player spawner working way better now
Replication uses component id to correctly place where the data of replicated variables are. Earlier it always assumed that they are always in the same spot and everyhting worked perfectly but usually did not. Its just generaly more stable.

+++ Eclipse-Source/Source/Engine/EclipsedEngine/EclipsedEngine/Reflection/AbstractSerializedVariable.h
@@ -63,7 +63,7 @@ namespace Eclipse::Reflection
 		};
 
 
-		std::unordered_map<ImGuiDataType, int> sizes = {
+		std::unordered_map<ImGuiDataType, size_t> sizes = {
 			{ImGuiDataType_S8, sizeof(char)},
 			{ImGuiDataType_U8, sizeof(unsigned char)},
